projects.api.test written; projects routes passing all tests
