+++ lib/routes/projects.js
@@ -4,39 +4,32 @@ const bodyParser = require('body-parser').json();
 const Project = require('../models/project');
 
 router
-	//get all
-	.get('/', (req, res, next) => {
+.get('/', (req, res, next) => {
   Project.find().lean()
-			.then(projects => res.send(projects))
-			.catch(next);
+    .then(projects => res.send(projects))
+    .catch(next);
 })
-
-//get by id
-    .get('/:id', (req, res, next) => {
-      Project.findById(req.params.id)
-					.then(project => res.send(project ))
-					.catch(next);
-    })
-
-	//delete by id
-	.delete('/:id', (req, res, next) => {
+.get('/:id', (req, res, next) => {
+  Project.findById(req.params.id)
+    .then(project => res.send(project ))
+    .catch(next);
+})
+.delete('/:id', (req, res, next) => {
   Project.findByIdAndRemove(req.params.id)
-			.then(deleted => res.send(deleted ))
-			.catch(next);
+    .then(deleted => res.send(deleted ))
+    .catch(next);
 })
-		
-	//post new project
-	.post('/', bodyParser, (req, res, next) => {
-  new Project(req.body).save()
-			.then(saved => res.send(saved))
-			.catch(next);
+.post('/', bodyParser, (req, res, next) => {
+  let proj = new Project(req.body);
+  proj.user = req.user.id;
+  proj.save()
+    .then(saved => res.send(saved))
+    .catch(next);
 })
-
-	//edit by id
-	.put('/:id', bodyParser, (req, res, next) => {
+.put('/:id', bodyParser, (req, res, next) => {
   Project.findByIdAndUpdate(req.params.id, req.body)
-            .then(saved => res.send(saved))
-            .catch(next);
+    .then(saved => res.send(saved))
+    .catch(next);
 });
 
 module.exports = router;
