+++ test/api-tests/users.api.test.js
@@ -5,9 +5,9 @@ chai.use(chaiHttp);
 
 process.env.MONGODB_URI = 'mongodb://localhost/bigtodo-test';
 
-const connection = require('../lib/setup-mongoose');
-const db = require('./db');
-const app = require('../lib/app');
+const connection = require('../../lib/setup-mongoose');
+const db = require('../db');
+const app = require('../../lib/app');
 
 describe ('Users API', () => {
